Create duplicate_translate_1dir.py
This script duplicates an origami structure from oxDNA and translates it along 1 direction. It takes an initial configuration and topology file for one origami (optionally, you can add a text file with indices from the original origami). It will return the initial configuration and topology for the new duplicated system consisting of two origamis, and the extended index selection for the new structure. It can be easily adapted to translate in 3 dimensions, or as needed.